Simplified matrix example and added a challenge
Hi Scott,

I think introducing the 'rnorm' function to generate values for a matrix could be confusing - especially if the students are trying to follow along and suddenly their command is producing different output from what you're showing on the screen! So I'm suggesting changing this to a simple sequence (which the students have already seen by that point). 

I also included a little challenge to encourage the students to consult the help page - so they start understanding there is often a way to change the default behaviour of a function.  

cheers -- Pip